+++ CHANGELOG.md
@@ -6,8 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- `--save` argument
-- `--file-format` argument
+- `--save-path` argument
+- `--save-format` argument
 ## [1.5] - 2025-02-26
 ### Added
 - `--info` argument